feat: allow initializing litellm guardrail with PII config
