+++ litellm/proxy/guardrails/guardrail_registry.py
@@ -9,6 +9,7 @@ from .guardrail_initializers import (
     initialize_guardrails_ai,
     initialize_hide_secrets,
     initialize_lakera,
+    initialize_litellm_pii,
     initialize_presidio,
 )
 
@@ -20,4 +21,5 @@ guardrail_registry = {
     SupportedGuardrailIntegrations.PRESIDIO.value: initialize_presidio,
     SupportedGuardrailIntegrations.HIDE_SECRETS.value: initialize_hide_secrets,
     SupportedGuardrailIntegrations.GURDRAILS_AI.value: initialize_guardrails_ai,
+    SupportedGuardrailIntegrations.LITELLM_PII.value: initialize_litellm_pii,
 }
